Changelog: - Update kernels to r111 and add kernel 6.1

+++ external-drivers/rtl8814au/remove-driver.sh
@@ -1,46 +1,97 @@
 #!/bin/bash
 
-SCRIPT_NAME="remove-driver.sh"
-SCRIPT_VERSION="20210917"
+# Purpose: Remove Realtek out-of-kernel USB WiFi adapter drivers.
+#
+# Supports dkms and non-dkms removals.
 
+SCRIPT_NAME="remove-driver.sh"
+SCRIPT_VERSION="20221018"
+MODULE_NAME="8814au"
 DRV_NAME="rtl8814au"
 DRV_VERSION="5.8.5.1"
-OPTIONS_FILE="8814au.conf"
+OPTIONS_FILE="${MODULE_NAME}.conf"
 
+KVER="$(uname -r)"
+KSRC="/lib/modules/${KVER}/build"
+MODDESTDIR="/lib/modules/${KVER}/kernel/drivers/net/wireless/"
+
+DRV_NAME="rtl${MODULE_NAME}"
+DRV_DIR="$(pwd)"
+
+# check to ensure sudo was used
 if [[ $EUID -ne 0 ]]
 then
 	echo "You must run this script with superuser (root) privileges."
-	echo "Try \"sudo ./${SCRIPT_NAME}\""
+	echo "Try: \"sudo ./${SCRIPT_NAME}\""
 	exit 1
 fi
 
-echo "Starting removal..."
+# support for the NoPrompt option allows non-interactive use of this script
+NO_PROMPT=0
+
+# get the script options
+while [ $# -gt 0 ]
+do
+	case $1 in
+		NoPrompt)
+			NO_PROMPT=1 ;;
+		*h|*help|*)
+			echo "Syntax $0 <NoPrompt>"
+			echo "       NoPrompt - noninteractive mode"
+			echo "       -h|--help - Show help"
+			exit 1
+			;;
+	esac
+	shift
+done
+
+# displays script name and version
+echo "Running ${SCRIPT_NAME} version ${SCRIPT_VERSION}"
+
+# check for and remove non-dkms installation
+if [[ -f "${MODDESTDIR}${MODULE_NAME}.ko" ]]
+then
+	echo "Removing a non-dkms installation."
+	rm -f ${MODDESTDIR}${MODULE_NAME}.ko
+	/sbin/depmod -a ${KVER}
+fi
+
+# information that helps with bug reports
+# kernel
+uname -r
+# architecture - for ARM: aarch64 = 64 bit, armv7l = 32 bit
+uname -m
+#getconf LONG_BIT (may be handy in the future)
 
 dkms remove -m ${DRV_NAME} -v ${DRV_VERSION} --all
 RESULT=$?
 
 # RESULT will be 3 if there are no instances of module to remove
-# however we still need to remove the files or the install script
-# will complain.
+# however we still need to remove various files or the install script
+# may complain.
 if [[ ("$RESULT" = "0")||("$RESULT" = "3") ]]
 then
-	echo "Deleting ${OPTIONS_FILE} from /etc/modprobe.d"
+	echo "Removing ${OPTIONS_FILE} from /etc/modprobe.d"
 	rm -f /etc/modprobe.d/${OPTIONS_FILE}
-	echo "Deleting source files from /usr/src/${DRV_NAME}-${DRV_VERSION}"
+	echo "Removing source files from /usr/src/${DRV_NAME}-${DRV_VERSION}"
 	rm -rf /usr/src/${DRV_NAME}-${DRV_VERSION}
 	echo "The driver was removed successfully."
-	echo "Info: You may now delete the driver directory if desired."
+	echo "You may now delete the driver directory if desired."
 else
 	echo "An error occurred. dkms remove error = ${RESULT}"
 	echo "Please report this error."
 	exit $RESULT
 fi
 
-read -p "Are you ready to reboot now? [y/N] " -n 1 -r
-echo    # move to a new line
-if [[ $REPLY =~ ^[Yy]$ ]]
+# if NoPrompt is not used, ask user some questions to complete removal
+if [ $NO_PROMPT -ne 1 ]
 then
-    reboot
+	read -p "Do you want to reboot now? (recommended) [y/N] " -n 1 -r
+	echo
+	if [[ $REPLY =~ ^[Yy]$ ]]
+	then
+		reboot
+	fi
 fi
 
 exit 0
